+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG for Sulu
 * dev-develop
     * BUGFIX      #2534 [ContactBundle]       Fixed static usage of media repository
     * ENHANCEMENT #2533 [RouteBundle]         Added locale to route-defaults
+    * BUGFIX      #2535 [PreviewBundle]       Wrapped website-kernel and append preview-specific configs
     * BUGFIX      #2530 [AdminBundle]         Included husky build which fixes the login translation issue
     * FEATURE     #2528 [AdminBundle]         Added form-abstraction for simple data-mapper forms
     * ENHANCEMENT #2526 [SearchBundle]        Introduced contexts for indexes to restrict selections